JZ: Fix failing test `test_nested_geo_paths_csv`
The test failed in the previous commit when we passed a `location_data`
dict with missing keys to `str.format()`.
To address this, I used `collections.defaultdict` to return empty
strings for missing key-values.
However, `str.format()` doesn't like seem to play ball with defaultdict
so I went the old school % operator which works well.

After solving that, I found that for for nested row dicts some sort
of traversal would need to be done to locate the appropriate nested
dicts where the geopoint field should be inserted. This was sorted
with the help of `reduce()`