Stream data at the form list endpoint based on settings.STREAM_DATA
The endpoint relies on Prefetch of some querysets, as a result we
have objects in memory when the query is executed and this seems
not to be cleared after the requests has been completed. Using the
streaming response seems to have less adverse effect and the memory
usage seems to be sane and freed up after a requests.