Ensure that >400 level HTTP errors are properly handled
Fixes: #1072

We discovored that the API endpoint to get an enketo edit url
encountered >400 level errors it did not properly set the HTTP
status code and instead returned the edit url as false.

This set of commits fixes that by raising an EnketoError exception
which has a friendly default error message that can be overriden.