Ensure that >= 400 level HTTP errors are properly handled (#1080)
* Ensure that >400 level HTTP errors are properly handled

Fixes: #1072

We discovored that the API endpoint to get an enketo edit url
encountered >400 level errors it did not properly set the HTTP
status code and instead returned the edit url as false.

This set of commits fixes that by raising an EnketoError exception
which has a friendly default error message that can be overriden.

* Move EnketoError exception to onadata.libs.exceptions.py

Moved this exception here for consistency..

* Report errors caught while trying to get enketo edit url

These are mostly 400 level errors.  We ask the user to contact support
and so keeping a record of the errors encountered is a good idea, for
referencing.

In the spirit of DRY, we merged the duplicated report_exception functions
in logger_tools and viewer_tools and moved it to common_tools.