Add management command apply_can_add_project_perms
Command apply_can_add_project_perms - applies can_add_project permission
to all users who have can_add_xform permission to a user/organization
profile.

This was necessary because we previously, before April 2018, did not
have can_add_project permission on the UserProfile and
OrganizationProfile classes. An attempt on doing this in migrations
seems not to be a recommended approach.