Add get_first_record() function.
Loop through instances instead of using queryset.first() method which
seems to be expensive because of the order_by id that it adds to the
query. This was observed on a table with millions of records on a
very fresh database, PostgreSQL selectect an IndexScan in the query
which was taking quite a long time execute and increased the number
of queries waiting to be processed.