Ensure `date_modified` is updated on Instance deletion via the API
Usually when an instance is deleted via the `set_deleted` instance
function. The `date_modified` is also set to the time the Instance is
last saved and this is then used to determine whether an export needs to
be re-generated. This change ensures that the API performs the same
functionality as the `set_deleted` function while deleting instances