fix bug NoneType object has no attribute push (#2403)
* fix bug NoneType object has no attribute push

the bug was raised by Celery and succeeded the bug ParsedInstance DoesNotExist. The ParsedInstance was being read immediately after being created which means it was not guranteed to be available in the replica. The fix aplied is to read the created ParsedInstance from the master database

* remove whitespace

* remove trailing whitespace

* fix formatting error

* do not run webhook signal as cellery task

also revert previous changes

* remove unnecessary quotes

* fix failing test

* read instance from master db when running signl