fix open-data endpoint timeout bug
there currently exists a peculiar intermittent bug where after ordering the queryset and
the first item is accessed such as instances[0] or by slicing instances[0:1] (as in the
the pagination implementation) the execution freezes and no result is returned.
This causes the server to timeout. The workaround only ensures we order and paginate
the results only when the queryset returns more than 1 item