Solve intermittent bug where form permissions are not applied for new forms (#2470)
* solve intermittent bug where form permissions are not applied

Django post_save signal is run at the end of of the Model.save() method. It is not guaranteed that the transaction will have commited. This leads to situation where XForm.DoesNotExist exception is at times raised when a new form is created

* refactor code

do not call set_project_perms_to_xform asynchronously is the post_save signal when a new form is created

* refactor code

* remove unnecessary CELERY_TASK_ALWAYS_EAGER override in tests

* fix failing tests

* refactor test case