use actual User, Group instances in migration file (#2655)
When running a migration with RunPython, the apps.get_model function provides historical models that are designed to match the state of the database at the point in time of the migration. Instances of FakeUser instead of the actual User model are used.

This change ensures we use the actual model instances and not historical model instances.

Move `RunPython` migration in a separate migration file to ensure schema changes are applied before custom Python code is executed.