fix Attachment.MultipleObjectsReturned if duplicate attachments is greater than 1
a bug in versions older than v5.2.0 allowed the same attachment to be saved more than once for duplicate Instances. This is because the TemporaryUploadedFile file object was included in a get_or_create lookup. TemporaryUploadedFile is not comparable to the value for field media_file stored in the DB